Reestructuración del codigo con los controllers
